[IMP] evaluation : formulas can return a format
This commit adds the possibility to define formats on formula functions.
So when rendering cells the format of the formula is returned if the
cell hasn't format.

To do that this commit allows when compiling a formula to define a
priority list of formats for the formula. The final formula format is
computed during evaluation.