[REF] sheet: move header visibility out of sheet plugin
This commit create a new core plugin `header_visibility.ts` that manage
the hidden cols/rows, and thus removing the management of hidden cols/rows
from `sheet.ts`.

This in addition to 1413e14e that moved the manegement of header size from
`sheet.ts` make it so the only information the Sheet plugin have to manage about
column rows is :
 - the number of cols/rows in the sheet
 - the mapping cellId/position and the inverse mapping position/cellId

Now instead of using `sheet.rows/cols.lenght` to get the size of the sheet,
we'll use the getters `getNumberCols/getNumberRows`.

Lastly, we can replace `Sheet` by `SheetId` in a lot of functions arguments
since we don't need the sheet object anymore.

Preparation to 2585911