[IMP] spreadsheet: Add data filters
Add data filters to o_spreadsheet. It is now possible to create tables in
sheets that allow to filter the rows that contain certain values.

Add a Filter plugin that manage the data of the filters and adapt their zone
with the grid modification. A new UI plugin FilterEvaluation will keep track
of which rows are currently filtered by the defined filters.

Split the HeaderVisibility plugin in a core plugin and an UI plugin.
The core plugin will manage the col/row hidden by the user, and the UI plugin
will combine these with the filtered rows. Many visibility getters are thus
moved in UIGetters and cannot be accessed in core plugins.

Commit 46e3795 changed adaptZone to not extend the zone if a new col/row was
added at the start of the zone. This commit make adaptZone have the same
behaviour if a new col/row was added at the end of the zone, the zone won't
be extended.

Add 2 new options to the sort command:

- hasNoHeader: disable the detection of header in the sorted zone

- emptyCellAsZero: empty cells are treated as number cells with value of 0

Moved the Sort items of the top bar menu to a new category "Data" that also
contains menu items to create/remove filters.

Odoo task ID : 2585911