[REF] popovers : improve popover component
The popovers components have some problems :
- It's currently not possible to define the height of a popover. The
"childHeight" props is only used for positioning.
- flipHorizontalOffset/flipVerticalOffset props are confusing to use.
They are sometime be positive, sometime negatives, and computed by hand
 to use popovers. Ideally we'd want the logic of it to be hidden in the component.

Refactored the popover component to now take as a props a Rectangle, and where it should be
displayed relative to this rectangle (at the right/bottom). The positioning logic is now inside the
component, which make it easier to use. Also allow the popover to either have a fixed size, or a
dynamic size (with CSS height or max-height).

Odoo task 2796248