[IMP] Charts: improve side-pannel clarity
Task Description:

This task aims to add some clarity in the invalid labels/
series ranges handling. The confirm button is now hide
when at least one range is invalid, and a new reset button
appears in the same case, allowing the user to revert the
ranges to the last valid ranges.

Related task(s)

Task-2877163