[FIX] figure: Change component structure to leverage navigator native behaviour
Up until now, we'd have to run some nauseating computations to display a
figure according to their position in the spreadsheet (both pane and
scroll of the panes were used). This commit introduc a new html
architecture where a figure is enclosed in a container that hide any
overflowing part. The figuer will then be translated inside that
container, effectively being hidden when positioned outside the visible
scrope of the wrapper.

This approach allows us to get rid of some of the more tedious
computations to determine which part of a figure should be visible as
well as offering the possibility to remove most of the borders logic as
well. All in all, simpler code.

In the long run, it'd be great to have the figure component completely
agnostic of its container and to have its parent (figure container) in
charge of handling the frozen pane/scroll.

Task 3103586