[IMP] images: delete images from the file store
Currently, images are never deleted from the file store server
when the corresponding image figure is deleted, bloating
the server more and more when adding and deleting images.

With this commit, images are garbage collected when we are sure they
won't be needed anymore.
We can't delete the image right when the user deletes an image because
he can always hit CTRL-Z to undo the change and make the image reappear.

Task 3093071