REF edition plugin:
the state is globally inconsistent. we sometimes reset some values,
sometimes we don't. functions need to be called in a specific order not
to explose everywhere (see
https://github.com/odoo/o-spreadsheet/pull/1739 disucssions)

This would allow to clear up things to update the state in a second part
to simplify the actual composer stuff, which is currently handled in the
spreadsheet component. I don't think it's wise to keep it that way
because it means it's only accessible through component interactions and
not through actions, which is an issue in another PR:
 https://github.com/odoo/o-spreadsheet/pull/1806