[FIX] Chart: handling multiple ! marks in range input doesn't cause errors
This commit fixes the problem of when the range input in the chart configuration panel contains multiple exclamation marks, an error will be thrown during the processing and cause the whole app freezing even crashed.

The basic idea is to reverse the splitted range string array first, and then get the XC and sheet name.

This commit does not consider how to correctly process multiple exclamation marks to get the correct sheet name. The resulting sheet name might be wrong.

task 3100211