[REF] evaluation: unify the reading of empty values
An evaluated formula can expect to be supplied with the reference
to a single cell or to a range of several cells.

Currently:
- when the formula expects a simple value, we provide the "null"
value when the reference points to an empty cell.
- when the formula expects a range of several cells, we provide
the "undifined" values ​​for the empty cells.

To improve typing and reading, this commit changes the second case
to also return "null" values ​​in cases of ranges with empty cells.